backend/requirements.txt
# Flask
Flask==3.1.0
Flask-CORS==5.0.1
Flask-SQLAlchemy==3.1.1

# LLM Providers via LangChain
langchain>=0.3.14
langchain-openai>=0.3.0
langchain-anthropic>=0.3.1
langchain-groq>=0.2.4
langchain-community>=0.3.14
langchain-core>=0.3.33

# Embeddings & Vector Store
sentence-transformers>=3.3.1
faiss-cpu>=1.13.0

# Document Processing
pypdf>=5.1.0

# Keyword Search
rank-bm25==0.2.2

# Database
SQLAlchemy>=2.0.36

# Observability
langfuse>=2.58.0

# Utilities
python-dotenv==1.0.1
marshmallow>=3.23.0
nltk>=3.9.1
